refactor(questions): extract aggregate-stats helpers into lib/questionStats
Move the 10 private helpers + 2 key constants + normalizeObjectifCMC
from questions.ts (1529 lines) into convex/lib/questionStats.ts.
Public API of questions.ts is unchanged; file drops to ~1272 lines.

A proper domain-based split (crud / search / stats) would have broken
every frontend caller of api.questions.*, so that's deferred.
